clip training 2 after fixing float tensor

+++ rl/actor.py
@@ -104,7 +104,7 @@ class Actor(nn.Module):
     task_string = line.strip().split(":")[0]
     task_string = eval(task_string)
     tokens = clip.tokenize(task_string).to("cuda")
-    text_features = self.clip_model.encode_text(tokens)
+    text_features = self.clip_model.encode_text(tokens).float()
     task_feat = F.relu(self.task_feat_block1(text_features))
     task_feat = F.relu(self.task_feat_block2(task_feat))
     task_feat = F.relu(self.task_feat_block3(task_feat))
